yt_app/requirements.txt
appnope==0.1.0
APScheduler==3.6.3
arrow==0.16.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1.1
django-crontab==0.7.1
djangorestframework==3.11.1
idna==2.10
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
q==2.6
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
traitlets==5.0.4
tzlocal==2.1
urllib3==1.25.10
wcwidth==0.2.5
